% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_area_between_KM_curves}
\alias{compute_area_between_KM_curves}
\title{Compute the Area Between Two Kaplan-Meier Curves}
\usage{
compute_area_between_KM_curves(start_time, end_time, Si, Sj)
}
\arguments{
\item{start_time}{Numeric, the start time of the integration interval.}

\item{end_time}{Numeric, the end time of the integration interval.}

\item{Si}{A Kaplan-Meier survival curve object/function for the first group.}

\item{Sj}{A Kaplan-Meier survival curve object/function for the second group.}
}
\value{
Numeric value representing the integrated absolute difference (area) between the two KM curves.
}
\description{
This function calculates the absolute area between two Kaplan-Meier survival curves
using numerical integration.
}
\examples{
\dontrun{
Si <- some_kaplan_meier_function(data1)
Sj <- some_kaplan_meier_function(data2)
area <- compute_area_between_KM_curves(0, 100, Si, Sj)
}

}
