% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Ri_nearest.R, R/nearestNeighbors.R
\name{get_Ri_nearest2}
\alias{get_Ri_nearest2}
\title{Get the indices of samples nearest to Ri}
\usage{
get_Ri_nearest2(dist.mat, Ri, nbd.method, k = 0, Ri.radius = NULL)

get_Ri_nearest2(dist.mat, Ri, nbd.method, k = 0, Ri.radius = NULL)
}
\arguments{
\item{dist.mat}{Distance matrix}

\item{Ri}{Sample Ri index.}

\item{nbd.method}{neighborhood method \code{"multisurf"} or \code{"surf"} (no k) 
or \code{"relieff"} (specify k). Used by nearestNeighbors().}

\item{k}{Integer for the number of neighbors (\code{"relieff"} method).}

\item{Ri.radius}{Radius of the neighborhood (other methods).}
}
\value{
Numeric vector of nearest indices.

Numeric vector of nearest indices.
}
\description{
Get the indices of samples nearest to Ri

Get the indices of samples nearest to Ri
}
