% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{objectiveFunction}
\alias{objectiveFunction}
\title{Cox Regression}
\usage{
objectiveFunction(beta, y, X)
}
\description{
This script defines an objective function for a Cox regression model and uses
the `optim` function to estimate the model's beta coefficients.
}
\examples{
# Define your data matrix X here
# X <- ...

# Initial guess for beta
beta_init <- rep(0, ncol(X))

# Use optim to find the beta that minimizes -L (maximizes L)
optim_result <- optim(par = beta_init, fn = objective_function, X = X, method = "BFGS")

# Extract the estimated beta coefficients
beta_est <- optim_result$par

}
\seealso{
\code{\link{optim}}
}
\author{
Your Name
}
\keyword{internal}
