% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Calculate the square of a number}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The square of x.
}
\description{
Calculate the square of a number
}
\examples{
square(4)
square(-3)
}
